requirements.txt
allure_python_commons==2.15.3
fastapi==0.133.1
httpx==0.28.1
passlib==1.7.4
pydantic==2.12.5
pydantic_settings==2.13.1
pytest==8.4.2
pytest_asyncio==1.2.0
python-dotenv==1.2.1
python_jose==3.5.0
SQLAlchemy==2.0.47
starlette==0.52.1
uvicorn==0.41.0
alembic>=1.13.0
asyncmy>=0.2.9
greenlet>=3.0.0
cryptography>=41.0.0
python-multipart>=0.0.6